added WebApi Project and fixed some repositories

+++ University.DataAccess/Repositories/AbstractAndInterfaces/BaseEntityRepository.cs
@@ -2,8 +2,8 @@
 {
     public abstract class BaseEntityRepository<T>(UniversityContext context) : BaseRepository<T>(context), ISingleIdMethods<T>
     {
-        public abstract Task<T?> GetByIdAsync(int id);
+        public abstract Task<T?> GetByIdAsync(long id);
 
-        public abstract Task DeleteByIdAsync(int id);
+        public abstract Task DeleteByIdAsync(long id);
     }
 }
