+++ University.DataAccess/Repositories/AbstractAndInterfaces/BaseThreeFkJoinTableRepository.cs
@@ -2,8 +2,8 @@
 {
     public abstract class BaseThreeFkJoinTableRepository<T>(UniversityContext context) : BaseRepository<T>(context), IThreeIdsMethods<T>
     {
-        public abstract Task<T?> GetByIdAsync(int id1, int id2, int id3);
+        public abstract Task<T?> GetByIdAsync(long id1, long id2, long id3);
 
-        public abstract Task DeleteByIdAsync(int id1, int id2, int id3);
+        public abstract Task DeleteByIdAsync(long id1, long id2, long id3);
     }
 }
