+++ University.DataAccess/Repositories/AbstractAndInterfaces/BaseTwoFkJoinTableRepository.cs
@@ -2,8 +2,8 @@
 {
     public abstract class BaseTwoFkJoinTableRepository<T>(UniversityContext context) : BaseRepository<T>(context), ITwoIdsMethods<T>
     {
-        public abstract Task<T?> GetByIdAsync(int id1, int id2);
+        public abstract Task<T?> GetByIdAsync(long id1, long id2);
 
-        public abstract Task DeleteByIdAsync(int id1, int id2);
+        public abstract Task DeleteByIdAsync(long id1, long id2);
     }
 }
